refactor: SearchBar input에 title 추가 및 디렉토리 위치 이동
- readOnly 속성에 따른 스타일 및 이동하는 경로 변경
