+++ frontend/src/components/SearchBar/SearchBar.styles.ts
